Refactored stats packages for better overview
