+++ src/main/java/dev/celestialfox/spectrumsurvival/utils/stats/impl/MongoDB.java
@@ -1,10 +1,11 @@
-package dev.celestialfox.spectrumsurvival.utils.stats;
+package dev.celestialfox.spectrumsurvival.utils.stats.impl;
 
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoClients;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.model.UpdateOptions;
+import dev.celestialfox.spectrumsurvival.utils.stats.IStatsStorage;
 import org.bson.Document;
 
 public class MongoDB implements IStatsStorage {
